RobotBase: shared base class with pickup/place/go_home/find_objects
All convenience methods (pickup, place, go_home, find_objects, holding,
get_object_pose, sim, abort) live in RobotBase. A new robot inherits
from it and only sets up arms + grasp_manager + named_poses.

_SimpleRobot in cli.py shrinks to ~20 LOC on top of RobotBase.
