+++ src/mj_manipulator/__init__.py
@@ -46,7 +46,7 @@ from mj_manipulator.protocols import (
     Gripper,
     IKSolver,
 )
-from mj_manipulator.robot import ManipulationRobot
+from mj_manipulator.robot import ManipulationRobot, RobotBase
 from mj_manipulator.sim_context import SimArmController, SimContext
 from mj_manipulator.trajectory import Trajectory, create_linear_trajectory
 
@@ -80,8 +80,9 @@ __all__ = [
     # Physics controller (simulation)
     "PhysicsController",
     "ArmPhysicsExecutor",
-    # Robot protocol
+    # Robot protocol and base class
     "ManipulationRobot",
+    "RobotBase",
     # Simulation context
     "SimContext",
     "SimArmController",
